♻️ refactor: 폴더명 _components로 변경
